requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-rest-framework==0.1.0
djangorestframework==3.10.3
httpie==1.0.3
idna==2.8
Pygments==2.4.2
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
